+++ src/files/files.controller.spec.ts
@@ -1,10 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { FilesController } from './files.controller';
-import {
-  STORAGE_SERVICE,
-  IStorageService,
-  StorageNamespace,
-} from '../storage/storage.interface';
+import { STORAGE_SERVICE, IStorageService } from '../storage/storage.interface';
 
 const mockStorageService: Partial<IStorageService> = {
   get: jest.fn().mockResolvedValue(null),